requirements.txt
pandas
numpy
scipy
apache-airflow
